requirements.txt
asgiref==3.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
defusedxml==0.7.1
dj-rest-auth==2.1.11
Django==3.2.8
django-allauth==0.45.0
django-extensions==3.1.3
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.2
MarkupSafe==2.0.1
modeltranslation==0.25
oauthlib==3.1.1
openapi-codec==1.3.2
packaging==21.2
pycparser==2.20
PyJWT==2.3.0
pyparsing==2.4.7
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
uuid==1.30